feat: enforce admin authentication inside format and user handler

+++ src/Functions/User/Handler.ts
@@ -1,11 +1,11 @@
 import {
-  APIGatewayProxyEventV2,
-  APIGatewayProxyHandlerV2,
+  APIGatewayProxyEventV2WithJWTAuthorizer,
+  APIGatewayProxyHandlerV2WithJWTAuthorizer,
   APIGatewayProxyResultV2,
 } from "aws-lambda";
 
+import { createResponse, getAuthContext } from "@/Functions/Common";
 import { User } from "@/Types";
-import { createResponse } from "../Common/Response";
 import {
   createUser,
   deleteUser,
@@ -14,8 +14,8 @@ import {
   updateUser,
 } from "./Service";
 
-export const handler: APIGatewayProxyHandlerV2 = async (
-  event: APIGatewayProxyEventV2,
+export const handler: APIGatewayProxyHandlerV2WithJWTAuthorizer = async (
+  event: APIGatewayProxyEventV2WithJWTAuthorizer,
 ): Promise<APIGatewayProxyResultV2> => {
   const { routeKey, pathParameters, body } = event;
   const id = pathParameters?.id ?? "";
@@ -44,14 +44,29 @@ export const handler: APIGatewayProxyHandlerV2 = async (
           return createResponse(400, { message: "missing request body" });
         }
 
+        const postContext = getAuthContext(event);
+        if (!postContext.isAdmin) {
+          return createResponse(403, { message: "forbidden" });
+        }
+
         const newUser: User = JSON.parse(body);
-        await createUser(newUser);
-        return createResponse(201, newUser);
+        const createdUser = await createUser(newUser);
+        return createResponse(201, createdUser);
 
       case "PUT /users/{id}":
         // update user
         if (!id || !body) {
           return createResponse(400, { message: "missing user id or body" });
+        }
+
+        const userToUpdate = await getUser(id);
+        if (!userToUpdate) {
+          return createResponse(404, { message: "user not found" });
+        }
+
+        const putContext = getAuthContext(event);
+        if (!putContext.isAdmin && userToUpdate?.id !== putContext.userId) {
+          return createResponse(403, { message: "forbidden" });
         }
 
         const updates = JSON.parse(body);
@@ -62,6 +77,11 @@ export const handler: APIGatewayProxyHandlerV2 = async (
         // delete user
         if (!id) {
           return createResponse(400, { message: "missing user id" });
+        }
+
+        const deleteContext = getAuthContext(event);
+        if (!deleteContext.isAdmin) {
+          return createResponse(403, { message: "forbidden" });
         }
 
         await deleteUser(id);
